rooming/requirements.txt
boto3==1.9.41
botocore==1.12.47
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-extensions==2.1.3
django-filter==2.0.0
django-livefield==3.1.0
django-minio-storage==0.2.2
django-model-utils==3.1.2
django-nose==1.4.6
django-storages==1.7.1
djangorestframework==3.8.2
docutils==0.14
future==0.17.1
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
line-bot-sdk==1.8.0
minio==4.0.6
newrelic==4.4.1.104
nose==1.3.7
Pillow==5.3.0
psycopg2-binary==2.7.5
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.7
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
xlrd==1.1.0
